backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
